+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ## Added
+- `pybricksdev.ble.nus` module.
 - `pybricksdev.ble.pybricks` module.
 
 ## [1.0.0-alpha.7] - 2021-05-17